fix. step3 피드백 반영 (map에서 다음에 랜덤으로 둘곳 찾도록 로직 수정)
